Google maps sdk added

+++ app/src/main/java/com/example/earthquake/Model.kt
@@ -1,4 +1,4 @@
 package com.example.earthquake
 data class EarthQuakeResponse(val earthquakes:List<EarthQuake>)
-public data class EarthQuake(val datetime: String, val depth: Float, val lng: Float, val src: String,
-                      val eqid: String,val magnitude: Float, val lat: Float)
+public data class EarthQuake(val datetime: String, val depth: Double, val lng: Double, val src: String,
+                      val eqid: String,val magnitude: Double, val lat: Double)
